requirements.txt
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
icalendar>=5.0.0,<6.0.0
flask>=3.0.0,<4.0.0
